[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14172 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,16 +3,15 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class Main {
+public class Main{
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
-        int num = (a+b)*(a-b);
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
+        long num = (a+b)*(a-b);
 
         System.out.println(num);
-
     }
 }
